src/requirements.txt
fastapi
pandas
psycopg2_binary
PyJWT
bcrypt
redis
streamlit
numpy